requirement.txt
pyyaml
jinja2
pycryptodome